+++ src/test/java/org/eclipse/keyple/distributed/integration/readerserverside/AsyncScenarioITest.java
@@ -11,68 +11,80 @@
  ************************************************************************************** */
 package org.eclipse.keyple.distributed.integration.readerserverside;
 
+import org.eclipse.keyple.core.service.PoolPlugin;
 import org.eclipse.keyple.core.service.SmartCardServiceProvider;
-import org.eclipse.keyple.distributed.LocalServiceServer;
-import org.eclipse.keyple.distributed.LocalServiceServerFactoryBuilder;
-import org.eclipse.keyple.distributed.RemotePluginClient;
-import org.eclipse.keyple.distributed.RemotePluginClientFactoryBuilder;
+import org.eclipse.keyple.distributed.*;
 import org.eclipse.keyple.distributed.integration.readerserverside.endpoint.StubAsyncEndpointClient;
 import org.eclipse.keyple.distributed.integration.readerserverside.endpoint.StubAsyncEndpointServer;
 import org.junit.After;
-import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TestName;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class AsyncScenarioITest extends BaseScenario {
 
-  private static final Logger logger = LoggerFactory.getLogger(AsyncScenarioITest.class);
-
-  @Rule public TestName testName = new TestName();
-
-  @Before
-  public void setUp() {
-
-    initLocalStubPlugin();
-
-    localServiceName = testName.getMethodName() + "_async";
-
-    StubAsyncEndpointServer endpointServer = new StubAsyncEndpointServer(localServiceName);
-    StubAsyncEndpointClient endpointClient = new StubAsyncEndpointClient(endpointServer);
-
-    localServiceExtension =
-        SmartCardServiceProvider.getService()
-            .registerDistributedLocalService(
-                LocalServiceServerFactoryBuilder.builder(localServiceName)
-                    .withAsyncNode(endpointServer)
-                    .build())
-            .getExtension(LocalServiceServer.class);
-
-    remotePluginClient =
-        SmartCardServiceProvider.getService()
-            .registerPlugin(
-                RemotePluginClientFactoryBuilder.builder(REMOTE_PLUGIN_NAME)
-                    .withAsyncNode(endpointClient, 2)
-                    .build())
-            .getExtension(RemotePluginClient.class);
-  }
+  private final StubAsyncEndpointServer endpointServer =
+      new StubAsyncEndpointServer(LOCAL_SERVICE_NAME);
+  private final StubAsyncEndpointClient endpointClient =
+      new StubAsyncEndpointClient(endpointServer);
 
   @After
   public void tearDown() {
     SmartCardServiceProvider.getService().unregisterPlugin(REMOTE_PLUGIN_NAME);
-    SmartCardServiceProvider.getService().unregisterDistributedLocalService(localServiceName);
+    SmartCardServiceProvider.getService().unregisterDistributedLocalService(LOCAL_SERVICE_NAME);
+    SmartCardServiceProvider.getService().unregisterPlugin(LOCAL_CARDRESOURCE_PLUGIN_NAME);
     SmartCardServiceProvider.getService().unregisterPlugin(LOCAL_PLUGIN_NAME);
   }
 
-  @Test
   @Override
-  public void execute_transaction_on_pool_reader() {
+  public void execute_transaction_with_regular_plugin() {
+
+    // Init server
+    initLocalStubPlugin();
+    SmartCardServiceProvider.getService()
+        .registerDistributedLocalService(
+            LocalServiceServerFactoryBuilder.builder(LOCAL_SERVICE_NAME)
+                .withAsyncNode(endpointServer)
+                .build())
+        .getExtension(LocalServiceServer.class);
+
+    // Init client
+    SmartCardServiceProvider.getService()
+        .registerPlugin(
+            RemotePluginClientFactoryBuilder.builder(REMOTE_PLUGIN_NAME)
+                .withAsyncNode(endpointClient, 2)
+                .build());
+
     try {
       Thread.sleep(1000);
     } catch (InterruptedException e) {
       e.printStackTrace();
     }
+  }
+
+  @Test
+  @Override
+  public void execute_transaction_with_pool_plugin() {
+
+    // Init server
+    initLocalStubPlugin();
+    initLocalCardResourceService();
+    initLocalCardResourcePlugin();
+    SmartCardServiceProvider.getService()
+        .registerDistributedLocalService(
+            LocalServiceServerFactoryBuilder.builder(LOCAL_SERVICE_NAME)
+                .withAsyncNode(endpointServer)
+                .withPoolPlugins(LOCAL_CARDRESOURCE_PLUGIN_NAME)
+                .build())
+        .getExtension(LocalServiceServer.class);
+
+    // Init client
+    remotePlugin =
+        (PoolPlugin)
+            SmartCardServiceProvider.getService()
+                .registerPlugin(
+                    RemotePoolPluginClientFactoryBuilder.builder(REMOTE_PLUGIN_NAME)
+                        .withAsyncNode(endpointClient, 2)
+                        .build());
+
+    executePoolPluginScenario();
   }
 }